simard-pay/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
airportsdata==20220406
astroid==2.9.0
async-timeout==4.0.1
attrdict==2.0.1
attrs==21.2.0
awsebcli==3.20.3
base58==2.1.1
bcrypt==3.2.2
bitarray==1.2.2
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cachelib==0.4.1
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
coverage==6.2
cryptography==37.0.1
cytoolz==0.11.2
Deprecated==1.2.13
distlib==0.3.4
dnspython==2.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.17.0
elasticsearch==7.15.2
elementpath==2.5.1
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.7.1
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
Flask-Testing==0.8.1
frozenlist==1.2.0
future==0.16.0
greenlet==1.1.2
hexbytes==0.2.2
identify==2.5.1
idna==3.3
importlib-metadata==4.8.2
ipfshttpclient==0.8.0a2
iso3166==2.0.2
iso4217==1.6.20180829
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lru-dict==1.1.7
MarkupSafe>=2.1.1
mccabe==0.6.1
mongomock==3.23.0
more-itertools==8.12.0
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
nodeenv==1.7.0
oauthlib==2.1.0
opentraveldata==0.0.9.post1
paramiko==2.10.4
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.19.0
protobuf==3.19.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pyflakes==2.4.0
pylint==2.12.2
pymongo==4.0
PyNaCl==1.5.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.1.0
rlp==2.0.1
semantic-version==2.8.5
sentinels==1.0.0
simplejson==3.17.6
six==1.14.0
SQLAlchemy==1.4.27
stripe==2.63.0
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
toolz==0.11.2
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.7
varint==1.0.2
virtualenv==20.15.0
wcwidth==0.1.9
web3==5.25.0
websocket-client==0.59.0
websockets==9.1
Werkzeug==2.2.3
wrapt==1.13.3
xmlschema==1.10.0
yarl==1.7.2
zipp==3.6.0
